refactor: fix copy order
