+++ src/services/items/constants/utils.ts
@@ -1,0 +1,16 @@
+import { Item } from '../interfaces/item';
+
+export const sortChildrenWith = (idsOrder: string[]) => (stElem: Item, ndElem: Item) => {
+  if (idsOrder.indexOf(stElem.id) >= 0 && idsOrder.indexOf(ndElem.id) >= 0) {
+    return idsOrder.indexOf(stElem.id) - idsOrder.indexOf(ndElem.id);
+  }
+  if (idsOrder.indexOf(stElem.id) >= 0) {
+    return -1;
+  }
+
+  if (idsOrder.indexOf(ndElem.id) >= 0) {
+    return 1;
+  }
+
+  return Date.parse(stElem.createdAt) - Date.parse(ndElem.createdAt);
+};
