+++ src/services/items/tasks/copy-item-task.ts
@@ -12,6 +12,8 @@ import { BaseItem } from '../base-item';
 import { Item } from '../interfaces/item';
 import { TaskStatus } from '../../..';
 import { FolderExtra } from './get-item-children-task';
+import { sortChildrenWith } from '../constants/utils';
+import { ITEM_TYPES } from '../constants/constants';
 
 type CopyItemSubTaskInput = { copy: Partial<Item>; original: Item; shouldCopyTags?: boolean };
 
@@ -144,12 +146,33 @@ export class CopyItemTask extends BaseItemTask<Item> {
 
   // replace children order with new ids
   private fixChildrenOrder(itemsMap: Map<string, { copy: Item; original: Item }>) {
+    const copyItemsArray = Array.from(itemsMap.values()).map(({ copy }) => copy);
     itemsMap.forEach((value) => {
       const { copy, original } = value;
-      if (original.extra?.folder) {
-        const { childrenOrder } = (original.extra as FolderExtra).folder;
-        const newOrder = childrenOrder.map((oldId) => itemsMap.get(oldId)?.copy?.id);
-        (copy.extra as FolderExtra).folder.childrenOrder = newOrder.filter(Boolean);
+      // set order for all copied folder
+      if (original.type === ITEM_TYPES.FOLDER) {
+        // init extra if necessary
+        if (!copy.extra.folder) {
+          copy.extra.folder = {};
+        }
+
+        const childrenOrder = (original.extra as FolderExtra)?.folder?.childrenOrder || [];
+
+        // change previous ids to copied item ids
+        const copyOrder = childrenOrder
+          .map((oldId) => itemsMap.get(oldId)?.copy?.id)
+          .filter(Boolean);
+
+        // get direct children
+        const children = copyItemsArray.filter(({ path }) => {
+          return path.startsWith(`${copy.path}.`);
+        });
+
+        // sort children to get wanter order -> get order by mapping to id
+        children.sort(sortChildrenWith(copyOrder));
+        const completeOrder = children.map(({ id }) => id);
+
+        (copy.extra as FolderExtra).folder.childrenOrder = completeOrder;
       }
 
       return value;
